refactor(admin): use Bun.file/stdin and shared flag parser in overview commands
Two simplifier-driven cleanups:

- Replace the manual readStdin Promise wrapper + node:fs/promises in
  overview-write with Bun.stdin.text() / Bun.file().text(), matching the
  pattern in onboard-apply.ts and add.ts.
- Extract parsePositiveIntFlag to src/lib/flags.ts and reuse from both
  overview-inputs and overview-write. Closes a validation gap on
  --release-count where Number(opts.releaseCount) silently coerced bad
  input to NaN.
